# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My swager Doc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "66fbe9538317e2bf92b72cfd76cd92dd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "name",
                                            "password",
                                            "phone_number"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Jhon"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "example@mail.ru"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "Qwerty"
                                            },
                                            "phone_number": {
                                                "type": "string",
                                                "example": "88005553535"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Jhon, registration complete."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "a13b421bcf9c9fb2a9b889f51bc4b856",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "phone_number",
                                            "password"
                                        ],
                                        "properties": {
                                            "phone_number": {
                                                "type": "string",
                                                "example": "88005553535"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "Qwerty"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "example": "3|5MhZkbe6xQXLQOSTJvwx24iZmywEuRxj5NGtkLRf610dc000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the following format: `Bearer {token}`",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}